// libs/ts-sdk/src/sdk/typegen/contracts/MiraAmmContract.ts
export type IdentityInput = Enum<{
  Address: AddressInput;
  ContractId: ContractIdInput;
}>;
export type IdentityOutput = Enum<{
  Address: AddressOutput;
  ContractId: ContractIdOutput;
}>;
// ...
export type StateOutput = Enum<{
  Uninitialized: void;
  Initialized: IdentityOutput;
  Revoked: void;
}>;

export type AddressInput = {bits: string};
export type AddressOutput = AddressInput;
export type AssetInput = {id: AssetIdInput; amount: BigNumberish};
export type AssetOutput = {id: AssetIdOutput; amount: BN};
export type AssetIdInput = {bits: string};
export type AssetIdOutput = AssetIdInput;
export type BurnEventInput = {
  pool_id: [AssetIdInput, AssetIdInput, boolean];
  recipient: IdentityInput;
  liquidity: AssetInput;
  asset_0_out: BigNumberish;
  asset_1_out: BigNumberish;
};
export type BurnEventOutput = {
  pool_id: [AssetIdOutput, AssetIdOutput, boolean];
  recipient: IdentityOutput;
  liquidity: AssetOutput;
  asset_0_out: BN;
  asset_1_out: BN;
};
export type ContractIdInput = {bits: string};
export type ContractIdOutput = ContractIdInput;
// ...
export type PoolMetadataOutput = {
  reserve_0: BN;
  reserve_1: BN;
  liquidity: AssetOutput;
  decimals_0: number;
  decimals_1: number;
};
// ...
const storageSlots: StorageSlot[] = [
  {
    key: "22bff5eb85bdd67aa101f965a7da5f5140424064c36d0a59097fef5439ff5675",
    value: "0000000000000000000000000000000000000000000000000000000000000000",
  },
  {
    key: "618f5159598624e8d2c389b09c5067df5674486eb10d0b37403686dfe080b17c",
    value: "0000000000000000000000000000000000000000000000000000000000000000",
  },
  {
    key: "618f5159598624e8d2c389b09c5067df5674486eb10d0b37403686dfe080b17d",
    value: "0000000000000000000000000000000000000000000000000000000000000000",
  },
  {
    key: "ea8bccdaf471150bc22cb2c940956b384c7404ddf20308253176c4ccb740113d",
    value: "0000000000000000000000000000000000000000000000000000000000000000",
  },
];
export class MiraAmmContractInterface extends Interface {
  constructor() {
    super(abi);
  }

  declare functions: {
    owner: FunctionFragment;
    decimals: FunctionFragment;
    name: FunctionFragment;
    symbol: FunctionFragment;
    total_assets: FunctionFragment;
    total_supply: FunctionFragment;
    burn: FunctionFragment;
    create_pool: FunctionFragment;
    fees: FunctionFragment;
    hook: FunctionFragment;
    mint: FunctionFragment;
    pool_metadata: FunctionFragment;
    set_hook: FunctionFragment;
    set_protocol_fees: FunctionFragment;
    swap: FunctionFragment;
    transfer_ownership: FunctionFragment;
  };
}

export class MiraAmmContract extends __Contract {
  static readonly abi = abi;
  static readonly storageSlots = storageSlots;

  declare interface: MiraAmmContractInterface;
  declare functions: {
    owner: InvokeFunction<[], StateOutput>;
    decimals: InvokeFunction<[asset: AssetIdInput], Option<number>>;
    name: InvokeFunction<[asset: AssetIdInput], Option<StdString>>;
    symbol: InvokeFunction<[asset: AssetIdInput], Option<StdString>>;
    total_assets: InvokeFunction<[], BN>;
    total_supply: InvokeFunction<[asset: AssetIdInput], Option<BN>>;
    burn: InvokeFunction<
      [pool_id: [AssetIdInput, AssetIdInput, boolean], to: IdentityInput],
      [BN, BN]
    >;
    create_pool: InvokeFunction<
      [
        token_0_contract_id: ContractIdInput,
        token_0_sub_id: string,
        token_1_contract_id: ContractIdInput,
        token_1_sub_id: string,
        is_stable: boolean,
      ],
      [AssetIdOutput, AssetIdOutput, boolean]
    >;
    fees: InvokeFunction<[], [BN, BN, BN, BN]>;
    hook: InvokeFunction<[], Option<ContractIdOutput>>;
    mint: InvokeFunction<
      [pool_id: [AssetIdInput, AssetIdInput, boolean], to: IdentityInput],
      AssetOutput
    >;
    pool_metadata: InvokeFunction<
      [pool_id: [AssetIdInput, AssetIdInput, boolean]],
      Option<PoolMetadataOutput>
    >;
    set_hook: InvokeFunction<[contract_id?: Option<ContractIdInput>], void>;
    set_protocol_fees: InvokeFunction<
      [volatile_fee: BigNumberish, stable_fee: BigNumberish],
      void
    >;
    swap: InvokeFunction<
      [
        pool_id: [AssetIdInput, AssetIdInput, boolean],
        asset_0_out: BigNumberish,
        asset_1_out: BigNumberish,
        to: IdentityInput,
        data?: Option<Bytes>,
      ],
      void
    >;
    transfer_ownership: InvokeFunction<[new_owner: IdentityInput], void>;
  };

  constructor(id: string | Address, accountOrProvider: Account | Provider) {
    super(id, abi, accountOrProvider);
  }
}